% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.shiny}
\alias{is.shiny}
\title{Is the calling environment a Shiny app}
\usage{
is.shiny(x)
}
\value{
logical indicating if the environment is a shiny domain or local session
}
\description{
\code{is.shiny}
}
\details{
Determine if the calling function or environment is an rstudio session
  or a shiny application.
}
\examples{

is.shiny()


library(shiny)

tapp <- function(){

  ui <- fluidPage(
    verbatimTextOutput("showme")
  )

  server <- function(input, output, session) {
    output$showme <- renderPrint({
      is.shiny()
    })
  }

  shinyApp(ui = ui, server = server, options = list(launch.browser = T))
}


}
\author{
Carl Boneri, \email{carl.boneri@whyles.com}
}
